reorganize lock testing
Shared backend logic goes into relevant testing module, leaving lock_unix and lock_windows to test platform specific locking behavior
